Fix unsafe int64→int narrowing for MinInputFiles
Use int64(len(manifests)) instead of int(config.MinInputFiles) to
avoid potential truncation on 32-bit platforms (CodeQL high severity).

+++ weed/plugin/worker/iceberg_maintenance_handler.go
@@ -847,7 +847,7 @@ func (h *IcebergMaintenanceHandler) rewriteManifests(
 		return "", fmt.Errorf("parse manifest list: %w", err)
 	}
 
-	if len(manifests) < int(config.MinInputFiles) {
+	if int64(len(manifests)) < config.MinInputFiles {
 		return fmt.Sprintf("only %d manifests, below threshold of %d", len(manifests), config.MinInputFiles), nil
 	}
